tes teroz

+++ app/controllers/LaporanController.php
@@ -1,4 +1,5 @@
 <?php
+ 
 
 class LaporanController extends \BaseController {
 
@@ -13,6 +14,7 @@ class LaporanController extends \BaseController {
 	}
 
 
+
 	/**
 	 * Show the form for creating a new resource.
 	 *
@@ -21,6 +23,9 @@ class LaporanController extends \BaseController {
 	public function create()
 	{
 		//
+
+		
+
 	}
 
 
@@ -33,17 +38,28 @@ class LaporanController extends \BaseController {
 	{
 
 		
-		$laporan = new Laporan();
-		$laporan->no_bukti = Input::get('nobukti');
-		$laporan->tanggal = Input::get('tanggal');
-		$laporan->keterangan  = Input::get('keterangan');
-		$laporan->no_perk    = Input::get('noperk');
-		$laporan->kredit  = Input::get('kredit');
-		$laporan->debet = Input::get('debet');
-		
-		$laporan->save();
 
-		 return Redirect::to('laporan');
+		if (Auth::check())
+		{
+			$laporan = new Laporan();
+			$laporan->no_bukti = Input::get('nobukti');
+			$laporan->tanggal = Input::get('tanggal');
+			$laporan->no_perk    = Input::get('noperk');
+			$laporan->kode_akun    = Input::get('akun');
+			$laporan->kredit  = Input::get('kredit');
+			$laporan->debet = Input::get('debet');
+			$laporan->keterangan  = Input::get('keterangan');
+			
+			$laporan->save();
+
+			 return Redirect::to('laporan')->with('pesan_simpan', 'Data Berhasil Disimpan');
+		}
+  		else{
+				 return Redirect::to('login')->with('belum_login', 'Anda Harus Login');
+			}
+
+			
+		
 	}
 
 
@@ -53,10 +69,17 @@ class LaporanController extends \BaseController {
 	 * @param  int  $id
 	 * @return Response
 	 */
-	public function show($id)
+	public function search()
 	{
-		//
-	}
+	$searchBy = Input::get('searchBy');
+    $kode_akun = Input::get('katakunci');
+    $searchResult = DB::table('laporan_kas_harian')->where($searchBy, 'LIKE', '%'.$kode_akun.'%')->get();
+
+    return View::make('listpencariandata')
+            ->with('searchBy', $searchBy)
+            ->with('kode_akun', $kode_akun)
+            ->with('searchResult', $searchResult);
+        }
 
 
 	/**
@@ -67,9 +90,17 @@ class LaporanController extends \BaseController {
 	 */
 	public function edit($id)
 	{
+
+
 		//
+		if (Auth::check())
+		{
 		$laporan = Laporan::find($id);
   		return View::make('editlaporan')->with('laporan', $laporan);
+  		}
+  		else{
+				 return Redirect::to('login')->with('belum_login', 'Anda Harus Login');
+			}
 
 	}
 
@@ -84,17 +115,21 @@ class LaporanController extends \BaseController {
 	{
 		
 		//
+		
   		$id   = Input::get('id');
   		$laporan = Laporan::find($id);
 
   		$laporan->no_bukti = Input::get('nobukti');
 		$laporan->tanggal = Input::get('tanggal');
-		$laporan->keterangan  = Input::get('keterangan');
 		$laporan->no_perk    = Input::get('noperk');
+		$laporan->kode_akun    = Input::get('akun');
 		$laporan->kredit  = Input::get('kredit');
 		$laporan->debet = Input::get('debet');
+		$laporan->keterangan  = Input::get('keterangan');
 		$laporan->save();
-		return Redirect::to('laporan');
+		return Redirect::to('laporan')->with('pesan_update', 'Data Update Berhasil');
+
+
 	}
 
 
@@ -106,10 +141,35 @@ class LaporanController extends \BaseController {
 	 */
 	public function delete($id)
 	{
-		//
+
+	if (Auth::check())
+	{
 	  $laporan = Laporan::find($id);
       $laporan->delete();
-      return Redirect::to('laporan');
+      return Redirect::to('laporan')->with('pesan_hapus', 'Data Berhasil dihapus');
+      	}
+  		else{
+				 return Redirect::to('login')->with('belum_login', 'Anda Harus Login');
+			}
+	}
+
+	public function export(){
+
+		
+	if (Auth::check())
+	{
+		  $laporan = Laporan::select('no_bukti', 'tanggal', 'keterangan','no_perk','kredit','debet')->get();
+		  Excel::create('Laporan Kas Harian', function($excel) use($laporan) {
+		    $excel->sheet('Sheet 1', function($sheet) use($laporan) {
+		        $sheet->fromArray($laporan);
+		    });
+		})->export('xlsx');
+	}
+  		else{
+				 return Redirect::to('login')->with('belum_login', 'Anda Harus Login');
+			}
+
+		
 	}
